+++ packages/common/src/types.ts
@@ -21,6 +21,9 @@ export type Heading = {
   level: number | 'index';
   group?: string;
   enumerator?: string;
+  // For external URLs
+  url?: string;
+  open_in_same_tab?: boolean;
 };
 
 export type SiteLoader = {
@@ -68,4 +71,22 @@ export type PageLoader = {
   footer?: FooterLinks;
   // This may not be defined
   dependencies?: Dependency[];
+};
+
+export type CommonTemplateOptions = {
+  favicon?: string;
+  logo?: string;
+  logo_dark?: string;
+  logo_text?: string;
+  logo_url?: string;
+  analytics_google?: string;
+  analytics_plausible?: string;
+  numbered_references?: boolean;
+  folders?: boolean;
+  style?: string;
+  hide_authors?: boolean;
+  hide_footer_links?: boolean;
+  hide_toc?: boolean;
+  hide_outline?: boolean;
+  outline_maxdepth?: number;
 };