+++ packages/site/src/components/Navigation/PrimarySidebar.tsx
@@ -7,6 +7,8 @@ import {
   useGridSystemProvider,
   useThemeTop,
   useIsWide,
+  useBaseurl,
+  withBaseurl,
 } from '@myst-theme/providers';
 import type { Heading } from '@myst-theme/common';
 import { Toc } from './TableOfContentsItems.js';
@@ -16,11 +18,12 @@ import * as Collapsible from '@radix-ui/react-collapsible';
 import { ChevronRightIcon } from '@heroicons/react/24/solid';
 
 export function SidebarNavItem({ item }: { item: SiteNavItem }) {
+  const baseurl = useBaseurl();
   if (!item.children?.length) {
     return (
       <ExternalOrInternalLink
         nav
-        to={item.url ?? ''}
+        to={withBaseurl(item.url, baseurl) ?? ''}
         className={classNames(
           'p-2 my-1 rounded-lg',
           'hover:bg-slate-300/30',
@@ -42,7 +45,7 @@ export function SidebarNavItem({ item }: { item: SiteNavItem }) {
       >
         <ExternalOrInternalLink
           nav
-          to={item.url ?? ''}
+          to={withBaseurl(item.url, baseurl) ?? ''}
           className={classNames('py-2 grow', {})}
           onClick={() => setOpen(!open)}
         >
@@ -66,7 +69,7 @@ export function SidebarNavItem({ item }: { item: SiteNavItem }) {
           <ExternalOrInternalLink
             nav
             key={action.url}
-            to={action.url || ''}
+            to={withBaseurl(action.url, baseurl) || ''}
             className={classNames(
               'p-2 my-1 rounded-lg',
               'hover:bg-slate-300/30',
@@ -84,7 +87,7 @@ export function SidebarNavItem({ item }: { item: SiteNavItem }) {
 export function SidebarNav({ nav }: { nav?: SiteManifest['nav'] }) {
   if (!nav) return null;
   return (
-    <div className="w-full px-1 dark:text-white">
+    <div className="w-full px-1 dark:text-white font-medium">
       {nav.map((item) => {
         return <SidebarNavItem key={'url' in item ? item.url : item.title} item={item} />;
       })}
@@ -126,12 +129,14 @@ export const PrimarySidebar = ({
   nav,
   footer,
   headings,
+  hide_toc,
   mobileOnly,
 }: {
   sidebarRef?: React.RefObject<HTMLElement>;
   nav?: SiteManifest['nav'];
   headings?: Heading[];
   footer?: React.ReactNode;
+  hide_toc?: boolean;
   mobileOnly?: boolean;
 }) => {
   const top = useThemeTop();
@@ -156,6 +161,7 @@ export const PrimarySidebar = ({
         'fixed',
         `xl:${grid}`, // for example, xl:article-grid
         'grid-gap xl:w-screen xl:pointer-events-none overflow-auto max-xl:min-w-[300px]',
+        { 'lg:hidden': nav && hide_toc },
         { hidden: !open, 'z-30': open, 'z-10': !open },
       )}
       style={{ top }}