+++ packages/site/src/components/Navigation/Search.tsx
@@ -174,7 +174,7 @@ function SearchShortcut() {
   return (
     <div
       aria-hidden
-      className="items-center hidden mx-1 font-mono text-sm text-gray-400 sm:flex gap-x-1"
+      className="items-center hidden mx-1 font-mono text-sm text-gray-600 sm:flex gap-x-1"
     >
       <kbd
         className={classNames(
@@ -506,8 +506,9 @@ function SearchForm({
     <>
       <form onSubmit={onSubmit}>
         <div className="relative flex w-full h-10 flow-row gap-x-1 ">
-          <label id={searchListID} htmlFor={searchInputID}>
-            <MagnifyingGlassIcon className="absolute text-gray-400 inset-y-0 start-0 h-10 w-10 p-2.5 aspect-square flex items-center pointer-events-none" />
+          <label id={searchLabelID} htmlFor={searchInputID}>
+            <MagnifyingGlassIcon className="absolute text-gray-600 inset-y-0 start-0 h-10 w-10 p-2.5 aspect-square flex items-center pointer-events-none" />
+            <span className="hidden">Search query</span>
           </label>
           <input
             autoComplete="off"
@@ -561,7 +562,7 @@ const SearchPlaceholderButton = forwardRef<
       {...props}
       className={classNames(
         className,
-        'flex items-center h-10 aspect-square sm:w-64 text-left text-gray-400',
+        'flex items-center h-10 aspect-square sm:w-64 text-left text-gray-600',
         'border border-gray-300 dark:border-gray-600',
         'rounded-lg bg-gray-50 dark:bg-gray-700',
         {