+++ themes/article/app/components/ArticlePage.tsx
@@ -25,7 +25,7 @@ export function ArticlePage({ article }: { article: PageLoader }) {
   const siteDesign: TemplateOptions = siteManifest?.options ?? {};
 
   const { projects } = siteManifest;
-  const { hide_footer_links, hide_outline, outline_maxdepth } = {
+  const { hide_footer_links, hide_outline, outline_maxdepth, hide_authors } = {
     ...siteDesign,
     ...pageDesign,
   };
@@ -45,7 +45,7 @@ export function ArticlePage({ article }: { article: PageLoader }) {
     >
       <BusyScopeProvider>
         <ExecuteScopeProvider enable={compute?.enabled ?? false} contents={article}>
-          <ArticleHeader frontmatter={project}>
+          <ArticleHeader frontmatter={project} hideAuthors={hide_authors}>
             <div className="pt-5 md:self-center h-fit lg:pt-0 col-body lg:col-margin-right-inset">
               <DownloadLinksArea />
               {compute?.enabled && compute.features.launchBinder && (
@@ -57,9 +57,9 @@ export function ArticlePage({ article }: { article: PageLoader }) {
               )}
             </div>
           </ArticleHeader>
-          <main
-            id="main"
-            className={classNames(grid, 'subgrid-gap col-screen', {
+          <article
+            data-name="article-page-main"
+            className={classNames('article', grid, 'subgrid-gap col-screen', {
               'pt-10': isIndex,
             })}
           >
@@ -67,12 +67,12 @@ export function ArticlePage({ article }: { article: PageLoader }) {
               <div className="flex items-center p-3 mb-10 border-y bg-slate-50 dark:bg-slate-600 border-y-slate-300 col-screen">
                 <Link
                   to={baseurl || '/'}
-                  className="flex gap-1 px-2 py-1 font-normal no-underline border rounded bg-slate-200 border-slate-600 hover:bg-slate-800 hover:text-white hover:border-transparent"
+                  className="flex gap-1 px-2 py-1 font-normal no-underline rounded border bg-slate-200 border-slate-600 hover:bg-slate-800 hover:text-white hover:border-transparent"
                 >
                   <ArrowLeftIcon
                     width="1rem"
                     height="1rem"
-                    className="self-center flex-none transition-transform group-hover:-translate-x-1"
+                    className="flex-none self-center transition-transform group-hover:-translate-x-1"
                   />
                   <span>Back to Article</span>
                 </Link>
@@ -84,12 +84,12 @@ export function ArticlePage({ article }: { article: PageLoader }) {
                 )}
                 <a
                   href={article.frontmatter?.exports?.[0]?.url}
-                  className="flex gap-1 px-2 py-1 font-normal no-underline border rounded bg-slate-200 border-slate-600 hover:bg-slate-800 hover:text-white hover:border-transparent"
+                  className="flex gap-1 px-2 py-1 font-normal no-underline rounded border bg-slate-200 border-slate-600 hover:bg-slate-800 hover:text-white hover:border-transparent"
                 >
                   <DocumentArrowDownIcon
                     width="1rem"
                     height="1rem"
-                    className="self-center flex-none transition-transform group-hover:-translate-x-1"
+                    className="flex-none self-center transition-transform group-hover:-translate-x-1"
                   />
                   <span>Download {article.kind}</span>
                 </a>
@@ -100,9 +100,10 @@ export function ArticlePage({ article }: { article: PageLoader }) {
               hideKeywords={!isIndex}
               hideTitle={isIndex}
               hideOutline={hide_outline}
+              hideAuthors={hide_authors}
               outlineMaxDepth={outline_maxdepth}
             />
-          </main>
+          </article>
           {!hide_footer_links && <FooterLinksBlock links={article.footer} />}
         </ExecuteScopeProvider>
       </BusyScopeProvider>