+++ packages/diagrams/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @myst-theme/diagrams
 
+## 0.17.1
+
+### Patch Changes
+
+- @myst-theme/providers@0.17.1
+
+## 0.17.0
+
+### Patch Changes
+
+- Updated dependencies [2049d12]
+  - @myst-theme/providers@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies [8fba231]
+  - @myst-theme/providers@0.16.0
+
 ## 0.15.2
 
 ### Patch Changes