+++ packages/search-minisearch/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @myst-theme/search-minisearch
 
+## 0.17.1
+
+### Patch Changes
+
+- @myst-theme/search@0.17.1
+
+## 0.17.0
+
+### Patch Changes
+
+- @myst-theme/search@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- @myst-theme/search@0.16.0
+
 ## 0.15.2
 
 ### Patch Changes