+++ themes/article/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @myst-theme/article
 
+## 0.17.1
+
+### Patch Changes
+
+- 4ee2d3f: Changes to DOM ordering to ensure that `main` encloses `article` and that there is only one `main` on a page in both themes.
+- Updated dependencies [b61c5ec]
+- Updated dependencies [630812f]
+- Updated dependencies [4ee2d3f]
+- Updated dependencies [8981edd]
+  - @myst-theme/site@0.17.1
+  - @myst-theme/providers@0.17.1
+  - @myst-theme/jupyter@0.17.1
+  - @myst-theme/styles@0.17.1
+  - @myst-theme/common@0.17.1
+  - @myst-theme/icons@0.17.1
+
+## 0.17.0
+
+### Minor Changes
+
+- e6921c2: Allow overriding URL that navigation bar logo links to.
+
+### Patch Changes
+
+- 2049d12: Upgrade to mystmd 1.5.1
+- Updated dependencies [e6921c2]
+- Updated dependencies [2049d12]
+- Updated dependencies [eb484d9]
+  - @myst-theme/site@0.17.0
+  - @myst-theme/providers@0.17.0
+  - @myst-theme/jupyter@0.17.0
+  - @myst-theme/common@0.17.0
+  - @myst-theme/styles@0.17.0
+  - @myst-theme/icons@0.17.0
+
+## 0.16.0
+
+### Patch Changes
+
+- Updated dependencies [8fba231]
+- Updated dependencies [208c87e]
+- Updated dependencies [d2415a1]
+- Updated dependencies [05945ec]
+- Updated dependencies [e2fa549]
+  - @myst-theme/providers@0.16.0
+  - @myst-theme/common@0.16.0
+  - @myst-theme/site@0.16.0
+  - @myst-theme/jupyter@0.16.0
+  - @myst-theme/styles@0.16.0
+  - @myst-theme/icons@0.16.0
+
 ## 0.15.2
 
 ### Patch Changes